requirements.txt
regex
pyinstaller
Pillow
